Update the edit user page so that an error message is given if someone tries to delete the primary admin user

+++ system/admin/edit/user.php
@@ -160,23 +160,39 @@ if ($del == "1") {
 }
 
 if ($really_del == "1") {
-
-	$sql = "DELETE FROM user_settings
-			WHERE user_id = '" . $uid . "'";
-	$result = mysql_query($sql,$connection);
-
-	$sql = "DELETE FROM update_data
-			WHERE user_id = '" . $uid . "'";
-	$result = mysql_query($sql,$connection);
-
-	$sql = "DELETE FROM users 
-			WHERE id = '" . $uid . "'";
-	$result = mysql_query($sql,$connection);
 	
-	$_SESSION['result_message'] = "User <font class=\"highlight\">" . $new_first_name . " " . $new_last_name . " (" . $new_username . ")</font> Deleted<BR>";
+	$sql = "SELECT id
+			FROM users
+			WHERE username = 'admin'";
+	$result = mysql_query($sql,$connection);
+	while ($row = mysql_fetch_object($result)) {
+		$temp_uid = $row->id;
+	}
 	
-	header("Location: ../users.php");
-	exit;
+	if ($temp_uid == $uid) {
+
+		$_SESSION['result_message'] = "The user <font class=\"highlight\">admin</font> cannot be deleted<BR>";
+
+	} else {
+
+		$sql = "DELETE FROM user_settings
+				WHERE user_id = '" . $uid . "'";
+		$result = mysql_query($sql,$connection);
+	
+		$sql = "DELETE FROM update_data
+				WHERE user_id = '" . $uid . "'";
+		$result = mysql_query($sql,$connection);
+	
+		$sql = "DELETE FROM users 
+				WHERE id = '" . $uid . "'";
+		$result = mysql_query($sql,$connection);
+		
+		$_SESSION['result_message'] = "User <font class=\"highlight\">" . $new_first_name . " " . $new_last_name . " (" . $new_username . ")</font> Deleted<BR>";
+		
+		header("Location: ../users.php");
+		exit;
+
+	}
 
 }
 ?>
